+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # Changelog
 
+## [4.0.0](https://github.com/google-github-actions/release-please-action/compare/v3.7.13...v4.0.0) (2023-12-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove most configuration options in favor of manifest configuration to configure the release-please-action
+* rewrite in typescript
+* remove command option in favor of setting `release-type` and `skip-github-release`/`skip-github-pull-request`
+* run on node20
+* **deps:** upgrade release-please to v16
+* v4 release
+
+### Features
+
+* **deps:** upgrade release-please to v16 ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+* remove command option in favor of setting `release-type` and `skip-github-release`/`skip-github-pull-request` ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+* remove most configuration options in favor of manifest configuration to configure the release-please-action ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+* run on node20 ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+* v4 release ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+
+
+### Bug Fixes
+
+* bump release-please from 16.3.0 to 16.3.1 ([#876](https://github.com/google-github-actions/release-please-action/issues/876)) ([c9291a9](https://github.com/google-github-actions/release-please-action/commit/c9291a968c291c9576e9e82a69ac316f60b980eb))
+* restore include-component-in-tag default (false) ([#888](https://github.com/google-github-actions/release-please-action/issues/888)) ([19431a1](https://github.com/google-github-actions/release-please-action/commit/19431a1126b0d092093511e4d9074add51d1e9e1))
+
+
+### Code Refactoring
+
+* rewrite in typescript ([9cf05a0](https://github.com/google-github-actions/release-please-action/commit/9cf05a09b133475347f729dc7d17b023081cc8d3))
+
+## [3.7.13](https://github.com/google-github-actions/release-please-action/compare/v3.7.12...v3.7.13) (2023-11-06)
+
+
+### Bug Fixes
+
+* **deps:** bump dependency versions ([#849](https://github.com/google-github-actions/release-please-action/issues/849)) ([9108012](https://github.com/google-github-actions/release-please-action/commit/910801217bf1ec07dd4367e66345f83918f7ae57))
+
 ## [3.7.12](https://github.com/google-github-actions/release-please-action/compare/v3.7.11...v3.7.12) (2023-09-28)